clean 1, fix examples
